fix(vm): preserve last start time during stats sync
